Introduce resolver
The resolver module implements a path resolving algorithm that
leverages node's own require.

This has a couple benefits:
* more likely to continue working as expected when features are
introduced to npm/node
* reduction of code duplication (npm/node already implements the module
resolution we are interested in for this project)
* allow the use of NODE_PATH variable to control module directory

Additionally, `resolver` leverages `Package` and `Import` constructors
which implement some of the same logic previously found on the `find`
function in `lib/importer.js`. The benefit of creating these
constructors will obvious when unit tests are introduced for this
project.

`Package`: implements logic related to retrieving package data, such as
getting the package.json, and getting the sass 'main' file for a given
package.

`Import`: implements logic to make sense of a sass import.