Inline non-local css files
If a module resolves its styles to a css file - as opposed to a sass file - then inline the css into the compiled output rather than adding an `import` statement into the css to attempt to resolve a local file.

This is achieved by passing node-sass the path without a file extension.